pyproject.toml
[project]
name = "gdocs-cli"
version = "0.1.0"
description = "A command-line interface for Google Docs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Talk-Point GmbH", email = "info@talk-point.de" }
]
keywords = ["google-docs", "cli", "document", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Office Suites",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.8.0",
]

[project.scripts]
gdocs = "gdocs_cli.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gdocs_cli"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (typer uses this)
    "B904",   # raise from in except - typer.Exit doesn't need chaining
    "SIM102", # nested if - sometimes clearer than combined with `and`
    "SIM105", # contextlib.suppress - try/except/pass can be clearer with comments
]

[tool.ruff.lint.isort]
known-first-party = ["gdocs_cli"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
